вынес функции по рендерингу в отд модуль, в функцию render добавил возможность принимать компоненты
